Refactor SerializerEmitter into stage-based emission
Split the monolithic 937-line SerializerEmitter into focused classes:
- EmitHelpers: shared value expressions, scalar classification, escaping
- FieldEmitter: scalar field layouts (one-line, line-breaks, list)
- CollectionEmitter: table and subsection-per-item rendering
- SerializerEmitter: coordinator (566 lines), type info, context, schema

Also extracted EmitPropertySerializations into EmitSectionProperty and
EmitNonSectionProperty for clarity. No behavioral changes.
